update build script
